Refactoring domain for projects (#1)
* Refactor ProjectView -> ProjectCardView with technologies instead of repos

* Introduce ProjectDetailsView

* Init old write entities

* Adding write entities #1

* Adding batch of write entities

* Almost done

* WIP

* Introduce ProjectDetailsView

* fix build

* Loading staging dump for ITs

* First exception handling impl

* Updating http status

* Refactoring exception management

* WIP GET project

* WIP GET project

* Working permissions

* Add colors in test logs

* Finish GET project endpoint

* Make it work with either ID or slug

* Permit access to swagger ui

---------

+++ application/rest-api-adapter/src/main/java/onlydust/com/marketplace/api/rest/api/adapter/authentication/hasura/HasuraJwtPayload.java
@@ -32,7 +32,7 @@ public class HasuraJwtPayload {
         @JsonProperty("x-hasura-allowed-roles")
         List<String> allowedRoles;
         @JsonProperty("x-hasura-githubUserId")
-        Integer githubUserId;
+        Long githubUserId;
         @JsonProperty("x-hasura-odAdmin")
         Boolean isAnOnlydustAdmin;
         @JsonProperty("x-hasura-user-id")
