+++ domain/src/main/java/onlydust/com/marketplace/api/domain/model/User.java
@@ -11,7 +11,7 @@ import java.util.UUID;
 public class User {
     UUID id;
     List<String> permissions;
-    Integer githubUserId;
+    Long githubUserId;
     String avatarUrl;
     String login;
 }
